+++ yesod-core/src/Yesod/Core/Internal/TH.hs
@@ -52,6 +52,7 @@ module Yesod.Core.Internal.TH
     , RouteAuthSpec(..)
     , roRouteAuth
     , setRouteAuthorization
+    , setRouteHandlerWrapper
     )
  where